requirements.txt
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
python-pptx>=0.6.21
python-docx>=0.8.11
openpyxl>=3.1.0
PyPDF2>=3.0.0
pytest>=7.0.0
